fix(admin): move day-lock UI from prevadzka-overview to dashboard
Scope correction — "Uzamknúť"/"Odomknúť" a s tým súvisiaci PDF/XLSX
export (#417/#419) mali byť pridané na /dashboard, nie
/prevadzka-overview. Presúva celý closed-day flow (stav, potvrdzovacie
dialógy, report-tasks export) na AdminDashboard.tsx vedľa jej pôvodného
(nezávislého) gramáž exportu. /prevadzka-overview je späť presne v
stave pred #417, vrátane pôvodného exportu "Dodanie podkladov".
