+++ frontend/src/pages/admin/PrevadzkaOverview.tsx
@@ -1,9 +1,8 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
-import { Check, AlertTriangle, X, Upload, Smartphone, FileText, FileSpreadsheet, Loader2, LockKeyhole, LockKeyholeOpen } from "lucide-react";
+import React, { useCallback, useEffect, useState } from "react";
+import { Check, AlertTriangle, X, Upload, Smartphone, FileText, FileSpreadsheet, Loader2 } from "lucide-react";
 import { useAuth } from "../../context/auth";
 import { useToast } from "../../context/ToastContext";
 import { logger } from "../../lib/logger";
-import ConfirmationModal from "../client/components/ui/ConfirmationModal";
 import { PageHead, Button, Card, Input } from "./ui";
 
 const API = import.meta.env.VITE_API_URL || "/api";
@@ -32,16 +31,6 @@ interface OverviewResponse {
   date: string;
   edupage: OverviewRow[];
   app: OverviewRow[];
-}
-
-interface ClosedDayResponse {
-  date: string;
-  is_closed: boolean;
-}
-
-interface ReportTaskResponse {
-  task_id: string;
-  status: "pending" | "processing" | "complete" | "failed";
 }
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
@@ -168,20 +157,12 @@ const CategoryCard: React.FC<{
 
 const PrevadzkaOverview: React.FC = () => {
   const { apiFetch } = useAuth();
-  const { error: toastError, success: toastSuccess } = useToast();
+  const { error: toastError } = useToast();
   const [date, setDate] = useState(() => toDateString(new Date()));
   const [data, setData] = useState<OverviewResponse | null>(null);
   const [loading, setLoading] = useState(false);
   const [xlsxLoading, setXlsxLoading] = useState(false);
   const [pdfLoading, setPdfLoading] = useState(false);
-  const [closedLoading, setClosedLoading] = useState(true);
-  const [isClosed, setIsClosed] = useState(false);
-  const [closeConfirmOpen, setCloseConfirmOpen] = useState(false);
-  const [unlockConfirmOpen, setUnlockConfirmOpen] = useState(false);
-  const [closing, setClosing] = useState(false);
-  const [unlocking, setUnlocking] = useState(false);
-  const [reportLoading, setReportLoading] = useState<"pdf" | "xlsx" | null>(null);
-  const closedRequestId = useRef(0);
 
   const handleExport = useCallback(
     async (fmt: "xlsx" | "pdf", setFmt: (v: boolean) => void) => {
@@ -230,122 +211,9 @@ const PrevadzkaOverview: React.FC = () => {
     }
   }, [apiFetch, date, toastError]);
 
-  const fetchClosedState = useCallback(async () => {
-    const requestId = ++closedRequestId.current;
-    setClosedLoading(true);
-    try {
-      const res = await apiFetch(`${API}/admin/closed-days/?date=${date}`);
-      if (!res.ok) throw new Error("closed-day status failed");
-      const payload = await res.json() as ClosedDayResponse;
-      if (requestId === closedRequestId.current) setIsClosed(payload.is_closed);
-    } catch (e) {
-      logger.error(e);
-      if (requestId === closedRequestId.current) {
-        setIsClosed(false);
-        toastError("Nepodarilo sa overiť, či je deň uzavretý.");
-      }
-    } finally {
-      if (requestId === closedRequestId.current) setClosedLoading(false);
-    }
-  }, [apiFetch, date, toastError]);
-
-  const handleCloseDay = useCallback(async () => {
-    setClosing(true);
-    try {
-      const res = await apiFetch(`${API}/admin/closed-days/`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ date }),
-      });
-      if (!res.ok) {
-        const body = await res.json().catch(() => ({}));
-        throw new Error(body?.error?.message || "close failed");
-      }
-      setIsClosed(true);
-      toastSuccess("Deň bol uzavretý.");
-    } catch (e) {
-      logger.error(e);
-      toastError(e instanceof Error && e.message !== "close failed" ? e.message : "Deň sa nepodarilo uzavrieť.");
-      await fetchClosedState();
-    } finally {
-      setClosing(false);
-    }
-  }, [apiFetch, date, fetchClosedState, toastError, toastSuccess]);
-
-  const handleUnlockDay = useCallback(async () => {
-    setUnlocking(true);
-    try {
-      const res = await apiFetch(`${API}/admin/closed-days/unlock/`, {
-        method: "DELETE",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ date }),
-      });
-      if (!res.ok) {
-        const body = await res.json().catch(() => ({}));
-        throw new Error(body?.error?.message || "unlock failed");
-      }
-      setIsClosed(false);
-      toastSuccess("Deň bol odomknutý.");
-    } catch (e) {
-      logger.error(e);
-      toastError(e instanceof Error && e.message !== "unlock failed" ? e.message : "Deň sa nepodarilo odomknúť.");
-      await fetchClosedState();
-    } finally {
-      setUnlocking(false);
-    }
-  }, [apiFetch, date, fetchClosedState, toastError, toastSuccess]);
-
-  const handleClosedDayExport = useCallback(async (fmt: "pdf" | "xlsx") => {
-    setReportLoading(fmt);
-    try {
-      const submit = await apiFetch(`${API}/admin/report-tasks/`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ date, format: fmt }),
-      });
-      if (!submit.ok) throw new Error("submit failed");
-      const submitted = await submit.json() as ReportTaskResponse;
-
-      let complete = false;
-      for (let attempt = 0; attempt < 60; attempt += 1) {
-        const statusRes = await apiFetch(`${API}/admin/report-tasks/${submitted.task_id}/`);
-        if (!statusRes.ok) throw new Error("poll failed");
-        const task = await statusRes.json() as ReportTaskResponse;
-        if (task.status === "failed") throw new Error("task failed");
-        if (task.status === "complete") {
-          complete = true;
-          break;
-        }
-        await new Promise((resolve) => window.setTimeout(resolve, 1000));
-      }
-      if (!complete) throw new Error("task timeout");
-
-      const download = await apiFetch(`${API}/admin/report-tasks/${submitted.task_id}/download/`);
-      if (!download.ok) throw new Error("download failed");
-      const blob = await download.blob();
-      const url = URL.createObjectURL(blob);
-      const anchor = document.createElement("a");
-      anchor.href = url;
-      anchor.download = `denny_prehlad_objednavok_${date}.${fmt}`;
-      document.body.appendChild(anchor);
-      anchor.click();
-      anchor.remove();
-      URL.revokeObjectURL(url);
-    } catch (e) {
-      logger.error(e);
-      toastError("Denný prehľad sa nepodarilo vygenerovať.");
-    } finally {
-      setReportLoading(null);
-    }
-  }, [apiFetch, date, toastError]);
-
   useEffect(() => {
     void fetchData();
   }, [fetchData]);
-
-  useEffect(() => {
-    void fetchClosedState();
-  }, [fetchClosedState]);
 
   return (
     <>
@@ -355,45 +223,13 @@ const PrevadzkaOverview: React.FC = () => {
         desc="Prehľad, ktoré prevádzky za daný deň dodali objednávky."
         actions={
           <>
-            <Input type="date" value={date} onChange={(e) => setDate(e.target.value)} disabled={closing || unlocking} style={{ width: "auto" }} />
+            <Input type="date" value={date} onChange={(e) => setDate(e.target.value)} style={{ width: "auto" }} />
             <Button variant="danger" onClick={() => handleExport("pdf", setPdfLoading)} disabled={pdfLoading || loading || !data}>
               {pdfLoading ? <Loader2 className="zpa-spin" /> : <FileText />} PDF
             </Button>
             <Button variant="primary" onClick={() => handleExport("xlsx", setXlsxLoading)} disabled={xlsxLoading || loading || !data}>
               {xlsxLoading ? <Loader2 className="zpa-spin" /> : <FileSpreadsheet />} XLSX
             </Button>
-            {!closedLoading && !isClosed && (
-              <Button variant="secondary" onClick={() => setCloseConfirmOpen(true)} disabled={closing}>
-                {closing ? <Loader2 className="zpa-spin" /> : <LockKeyhole />} Uzamknúť
-              </Button>
-            )}
-            {!closedLoading && isClosed && (
-              <>
-                <span role="status" style={{ color: "var(--green-700)", fontWeight: 700, whiteSpace: "nowrap" }}>
-                  <Check style={{ width: 16, verticalAlign: "middle", marginRight: 5 }} />
-                  Deň je uzavretý
-                </span>
-                <Button
-                  variant="danger"
-                  aria-label="PDF objednávok"
-                  onClick={() => handleClosedDayExport("pdf")}
-                  disabled={reportLoading !== null}
-                >
-                  {reportLoading === "pdf" ? <Loader2 className="zpa-spin" /> : <FileText />} PDF objednávok
-                </Button>
-                <Button
-                  variant="primary"
-                  aria-label="XLSX objednávok"
-                  onClick={() => handleClosedDayExport("xlsx")}
-                  disabled={reportLoading !== null}
-                >
-                  {reportLoading === "xlsx" ? <Loader2 className="zpa-spin" /> : <FileSpreadsheet />} XLSX objednávok
-                </Button>
-                <Button variant="secondary" onClick={() => setUnlockConfirmOpen(true)} disabled={unlocking}>
-                  {unlocking ? <Loader2 className="zpa-spin" /> : <LockKeyholeOpen />} Odomknúť
-                </Button>
-              </>
-            )}
           </>
         }
       />
@@ -406,25 +242,6 @@ const PrevadzkaOverview: React.FC = () => {
           <CategoryCard title="App prevádzky" icon={<Smartphone />} rows={data?.app ?? []} source="app" />
         </div>
       )}
-
-      <ConfirmationModal
-        isOpen={closeConfirmOpen}
-        onClose={() => setCloseConfirmOpen(false)}
-        onConfirm={() => void handleCloseDay()}
-        title="Uzamknúť objednávky na tento deň?"
-        description="Po uzamknutí už nebude možné upravovať objednávky pre tento deň. Prípadné odomknutie bude vyžadovať samostatné potvrdenie."
-        confirmText="Uzamknúť"
-        variant="warning"
-      />
-      <ConfirmationModal
-        isOpen={unlockConfirmOpen}
-        onClose={() => setUnlockConfirmOpen(false)}
-        onConfirm={() => void handleUnlockDay()}
-        title="Odomknúť objednávky na tento deň?"
-        description="Odomknutím sa deň znova otvorí na úpravy objednávok, diét a ďalších údajov."
-        confirmText="Odomknúť"
-        variant="warning"
-      />
     </>
   );
 };
